requirements.txt
PyAudio
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-groq==0.0.1
langchain-openai==0.0.7
langsmith==0.1.6
faiss-cpu
langchain-community
PyPDF2
sendgrid
python-dotenv~=1.0.1
salesgpt~=0.1.2
pydantic~=2.7.3
certifi~=2024.2.2
requests~=2.32.3
deepgram-sdk
